+++ src/widgets/grid_widget.rs
@@ -29,6 +29,7 @@ use sdl2::pixels::Color;
 use sdl2::rect::{Point, Rect};
 use std::any::Any;
 use std::collections::HashMap;
+use crate::render::texture_cache::TextureCache;
 
 /// This is the storage object for the `GridWidget`.  It stores the config, properties, callback registry.
 pub struct GridWidget {
@@ -72,7 +73,7 @@ impl GridWidget {
 /// This is the `Widget` implementation of the `GridWidget`.
 impl Widget for GridWidget {
     /// Draws the `GridWidget` contents.
-    fn draw(&mut self, c: &mut Canvas<Window>) -> Option<&Texture> {
+    fn draw(&mut self, c: &mut Canvas<Window>, _t: &mut TextureCache) -> Option<&Texture> {
         if self.get_config().invalidated() {
             let bounds = self.get_config().get_size(CONFIG_SIZE);