+++ src/widgets/list_widget.rs
@@ -30,6 +30,7 @@ use sdl2::rect::Rect;
 use std::any::Any;
 use std::collections::HashMap;
 use std::path::Path;
+use crate::render::texture_cache::TextureCache;
 
 /// This is the callback type that is used when an `on_selected` callback is triggered from this
 /// `Widget`.
@@ -182,7 +183,7 @@ impl CanvasHelper for ListWidget {}
 /// This is the `Widget` implementation of the `ListWidget`.
 impl Widget for ListWidget {
     /// Draws the `ListWidget` contents.
-    fn draw(&mut self, c: &mut Canvas<Window>) -> Option<&Texture> {
+    fn draw(&mut self, c: &mut Canvas<Window>, _t: &mut TextureCache) -> Option<&Texture> {
         let base_color = self.get_color(CONFIG_COLOR_BASE);
 
         c.set_draw_color(base_color);