+++ src/widgets/text_widget.rs
@@ -29,6 +29,7 @@ use sdl2::rect::Rect;
 use std::any::Any;
 use std::collections::HashMap;
 use std::path::Path;
+use crate::render::texture_cache::TextureCache;
 
 /// This enum is used by the `TextWidget`, which controls the justification of the text being
 /// rendered within the bounds of the `Widget`.
@@ -101,7 +102,7 @@ impl TextWidget {
 /// copied to the canvas after rendering.  It uses blended mode texture mapping, which may be slow (as
 /// described by the SDL2 documentation), so this might change later to use 8 bit color mapping.
 impl Widget for TextWidget {
-    fn draw(&mut self, c: &mut Canvas<Window>) -> Option<&Texture> {
+    fn draw(&mut self, c: &mut Canvas<Window>, _t: &mut TextureCache) -> Option<&Texture> {
         let base_color = self.get_color(CONFIG_COLOR_BASE);
         let text_max_width =
             self.get_size(CONFIG_SIZE)[0] - ((self.get_numeric(CONFIG_BORDER_WIDTH) * 2) as u32);